Fixed stream parameter appearing twice in TTT

+++ devops/devops.py
@@ -4,12 +4,13 @@ from dotenv import load_dotenv
 load_dotenv()
 
 HELP_TEXT = """
-1) build                            ; This will build the docker image only
-2) start                            ; This will build the docker image and start the container
-3) stop                             ; This will stop the container
-4) logs                             ; This iwll show the container logs
+1) build                            ; This will build the docker image only based on one of ['ttt','stt','tts','itt']
+2) start                            ; This will build the docker image and start the container based on one of ['ttt','stt','tts','itt']
+3) stop                             ; This will stop the container based on one of ['ttt','stt','tts','itt']
+4) logs                             ; This iwll show the container logs based on one of ['ttt','stt','tts','itt']
 5) ps                               ; This will show the container status
-6) build_mistral7b_exllama
+6) push                             ; This will build and push the docker image based on one of ['ttt','stt','tts','itt']
+7) publish                          ; This will package distribution and publish to pypi based on one of ['ttt','stt','tts','itt']
 """
 
 class Deploy(cmd.Cmd):
@@ -24,29 +25,50 @@ class Deploy(cmd.Cmd):
         self._cmd(f"docker exec -it gai-{svc} bash")
 
     def do_ssh(self,svc):
+        if not svc:
+            print("Please specify one from ['ttt','stt','tts','itt'] ")
+            return
         self._ssh(svc)
 
     def do_logs(self,svc):
+        if not svc:
+            print("Please specify one from ['ttt','stt','tts','itt'] ")
+            return
         self._cmd(f"docker logs gai-{svc}")
 
     def do_exit(self,ignored):
         return True
 
+    def get_version(self):
+        with open('../setup.py', 'r') as file:
+            first_line = file.readline()
+            _, version = first_line.split('=')
+            version = version.strip().strip("'")  # remove whitespace and quotes
+        return version
+
     def do_build(self,svc):
+        if not svc:
+            print("Please specify one from ['ttt','stt','tts','itt'] ")
+            return
+        version =self.get_version()
         if (svc != "itt"):
             self._cmd("rm -rf working && mkdir working")
             self._cmd("cp ../gai.json working")
             self._cmd("cp -rp ../gai/gen/api working")
-            self._cmd(f"docker build -t gai-{svc}:latest -f Dockerfiles/Dockerfile.{svc.upper()} .")
+            self._cmd(f"docker build -t gai-{svc}:{version} -f Dockerfiles/Dockerfile.{svc.upper()} .")
         else:
             self._cmd("rm -rf working && mkdir working")
             self._cmd("cp ../gai.json working")
             self._cmd("cp -rp ../gai/gen/api working")
             self._cmd("cp -rp ../external/LLaVA working")
-            self._cmd(f"docker build -t gai-{svc}:latest -f Dockerfiles/Dockerfile.{svc.upper()} .")
+            self._cmd(f"docker build -t gai-{svc}:{version} -f Dockerfiles/Dockerfile.{svc.upper()} .")
 
     def do_start(self,svc):
+        if not svc:
+            print("Please specify one from ['ttt','stt','tts','itt'] ")
+            return
         self.do_stop(svc)
+        self.do_build(svc)
         self._cmd(f"""docker run -d \
             --name gai-{svc} \
             -p 12031:12031 \
@@ -59,16 +81,27 @@ class Deploy(cmd.Cmd):
             """)
 
     def do_stop(self,svc):
+        if not svc:
+            print("Please specify one from ['ttt','stt','tts','itt'] ")
+            return
         self._cmd(f"""docker rm -f gai-{svc}""")
 
     def do_push(self,svc):
-        self._cmd(f"""docker tag gai-{svc}:latest kakkoii1337/gai-{svc}:latest""")        
-        self._cmd(f"""docker push kakkoii1337/gai-{svc}:latest""")
+        if not svc:
+            print("Please specify one from ['ttt','stt','tts','itt'] ")
+            return
+        version=self.get_version()
+        self.do_build(svc)
+        self._cmd(f"""docker tag gai-{svc}:{version} kakkoii1337/gai-{svc}:{version}""")        
+        self._cmd(f"""docker push kakkoii1337/gai-{svc}:{version}""")
 
-    def do_ps(self,svc):
+    def do_ps(self,ignored):
         self._cmd(f"""docker ps -a""")
 
-    def do_publish_gai(self, ignored):
+    def do_publish(self, svc):
+        if not svc:
+            print("Please specify one from ['ttt','stt','tts','itt'] ")
+            return
         self._cmd("cd ~/github/kakkoii1337/gai-gen && python setup.py sdist")
         self._cmd("cd ~/github/kakkoii1337/gai-gen && twine upload dist/*")
